requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==4.2
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
phonenumbers==9.0.5
pillow==11.2.1
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.3
urllib3==2.4.0
whitenoise==6.9.0
